+++ bin/index.ts
@@ -1,18 +1,33 @@
+/*!
+ * © 2023 Atypon Systems LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import { promises as fs } from 'fs';
 import * as path from 'path';
 
+import { generateLookup } from './generateLookup';
+import { generateTypes } from './generateTypes';
 import { generateValidators } from './generateValidators';
-import { generateDefinitions } from './generateDefinitions';
-import { generateTypeLookups } from './generateTypeLookups';
 
 async function main() {
-    await fs.mkdir('dist', { recursive: true });
-    await fs.mkdir(path.join('dist', 'cjs'));
-    await fs.mkdir(path.join('dist', 'es'));
-    await fs.mkdir(path.join('dist', 'types'));
-    await generateValidators();
-    await generateDefinitions();
-    await generateTypeLookups();
+  await fs.mkdir('dist', { recursive: true });
+  await fs.mkdir(path.join('dist', 'cjs'));
+  await fs.mkdir(path.join('dist', 'es'));
+  await fs.mkdir(path.join('dist', 'types'));
+  await generateValidators();
+  await generateTypes();
+  await generateLookup();
 }
 
-main().catch(e => console.log(e));
+main().catch((e) => console.log(e));
