+++ package.json
@@ -16,7 +16,7 @@
     "build:es": "tsc --outDir dist/es --declarationDir dist/types --declaration",
     "generate:build": "tsc --project bin/tsconfig.json",
     "generate:run": "node tsc-dist/index.js",
-    "lint": "eslint src tests --ext .js,.ts",
+    "lint": "eslint src bin --ext .js,.ts",
     "prebuild": "rimraf tsc-dist/* dist/* schemas/* definitions/*",
     "test": "jest --runInBand",
     "typecheck": "tsc --noEmit",
@@ -27,15 +27,24 @@
     "deepmerge": "^4.3.1"
   },
   "devDependencies": {
+    "@manuscripts/examples": "^0.1.1",
+    "@manuscripts/eslint-config": "^0.5.1",
     "@typescript-eslint/eslint-plugin": "^5.47.0",
     "@typescript-eslint/parser": "^5.47.0",
+    "@types/jest": "^29.2.4",
     "eslint": "^8.30.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-header": "^3.1.1",
     "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^27.1.7",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-mdx": "^2.0.5",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-simple-import-sort": "^8.0.0",
     "jest": "^29.3.1",
     "json-schema-to-typescript": "^12.0.0",
     "npm-run-all": "^4.1.5",
