+++ src/index.ts
@@ -1,5 +1,20 @@
-export * from './validate'
-//@ts-ignore
-export * from './lookup'
-//@ts-ignore
-export * from './types'
+/*!
+ * © 2023 Atypon Systems LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+export * from './validate';
+// see bin/generateLookup.ts for how this file is generated.
+export * from './lookup';
+// see bin/generateTypes.ts for how this file is generated.
+export * from './types';
